update arena_battle.py
Feature is completely untested.
If it works, the top two links on youtube should be found and kept in variables the code "link1" and "link2". They are not displayed on the website because I did not want to override jj's code. Also, if there are no valid search results, there will be an error.